+++ packages/keystore/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @thorswap-lib/keystore
 
+## 1.0.0-development.143
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @thorswap-lib/toolbox-cosmos@0.1.0-development.90
+
 ## 1.0.0-development.142
 
 ### Patch Changes